requirements.txt
Django==4.1.2
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2>=2.8
